Fix some type inference problems

+++ src/core.jl
@@ -691,12 +691,12 @@ colordim{C<:ColorType}(img::AbstractMatrix{C}) = 0
 colordim{C<:ColorType}(img::AbstractArray{C,3}) = 0
 colordim{C<:ColorType}(img::AbstractImage{C}) = 0
 colordim(img::AbstractMatrix) = 0
-colordim{T}(img::AbstractImageDirect{T,3}) = get(img, "colordim", 0)
+colordim{T}(img::AbstractImageDirect{T,3}) = get(img, "colordim", 0)::Int
 colordim{T}(img::AbstractArray{T,3}) = (size(img, defaultarraycolordim) == 3) ? 3 : 0
 colordim(img::AbstractImageDirect) = get(img, "colordim", 0)
 colordim(img::AbstractImageIndexed) = 0
 
-timedim(img) = get(img, "timedim", 0)
+timedim(img) = get(img, "timedim", 0)::Int
 
 limits(img::AbstractArray{Bool}) = 0,1
 # limits{T<:Integer}(img::AbstractArray{T}) = typemin(T), typemax(T)  # best not to use Integers...
